# Bench triangle