--- a/include/intersectables/triangle.h
+++ b/include/intersectables/triangle.h
@@ -21,7 +21,7 @@ class Triangle : public Intersectable {
     Point3f* b;
     Point3f* c;
 
-    virtual Point3f* computeNormal(float alpha, float beta);
+    virtual Point3f* computeNormal(float alpha, float beta) const;
 
     Triangle(int, Material*, Point3f*, Point3f*, Point3f*);
     // plane-ray intersection ray: p(t) = orig + t * dir
@@ -30,8 +30,8 @@ class Triangle : public Intersectable {
     // f(p) = dot(n, (p - a)) intersection:
     // f(p(t)) = 0. Solve for t.  Plug t_i into
     // p(t_i) will give intersection point
-    HitRecord* intersect(Ray* ray);
-    virtual void log() {
+    HitRecord* intersect(Ray* ray) const;
+    virtual void log() const {
       std::cout << "face-index: " << faceId << std::endl;
       std::cout << "vertices: " <<  std::endl;
       a->log();