--- a/src/intersectables/intersectableList.cpp
+++ b/src/intersectables/intersectableList.cpp
@@ -6,7 +6,7 @@ IntersectableList::IntersectableList() {
   container = new std::vector<Intersectable*>;
 }
 
-HitRecord* IntersectableList::intersect(Ray* ray) {
+HitRecord* IntersectableList::intersect(Ray* ray) const {
   float minT = std::numeric_limits<float>::max();
   auto hitRecord = new HitRecord();
   for (auto const& intersectable: *container) {