Improves performance of triangle intersection. On my machine:

1435ns -> 1045 ns