@simplay 2019-03-04
Thanks for this contribution. :smiley_cat: I can confirm that this patch improves the overall performance. On my machine, I get the following timings:

## Running the benchmark:

**before:**

BM_TriangleIntersect        990 ns          990 ns       678425

**after**:

BM_TriangleIntersect        715 ns          715 ns       872831

## Running hyperfine:

**before**:
`~/repos/cpp-project/build (55349a7) ✔ hyperfine --warmup 2 --min-runs 2 "./main.x 200 200 7 2"`

```
Benchmark #1: ./main.x 200 200 7 2
Time (mean ± σ):      2.589 s ±  0.111 s    [User: 28.518 s, System: 0.047 s]
Range (min … max):    2.511 s …  2.668 s
```

**after**:

`~/repos/cpp-project/build (7874f9e) ✔ hyperfine --warmup 2 --min-runs 2 "./main.x 200 200 7 2"`

```
Benchmark #1: ./main.x 200 200 7 2
Time (mean ± σ):      1.811 s ±  0.007 s    [User: 20.159 s, System: 0.026 s]
Range (min … max):    1.806 s …  1.816 s
```

For sanity's sake, I re-rendered all test-scenes and they still looked correct to me.

One tiny request: Would you mind rebasing this patch against the latest master branch?

Apart from that, this PR LGTM :smiley_cat: :+1: